Move DependencyGraphTranspose and AdjointCache classes
